DX-1208: WIP commit for examples refactor into their own independent examples
