+++ imx/examples/order/go.mod
@@ -1,12 +1,14 @@
-module github.com/immutable/imx-core-sdk-golang/imx/examples
+module github.com/immutable/imx-core-sdk-golang/imx/examples/order
 
 go 1.18
 
-replace github.com/immutable/imx-core-sdk-golang => ./../../
+replace github.com/immutable/imx-core-sdk-golang => ./../../../
+
+replace github.com/immutable/imx-core-sdk-golang/imx/examples => ./../
 
 require (
 	github.com/immutable/imx-core-sdk-golang v0.0.0-00010101000000-000000000000
-	github.com/joho/godotenv v1.4.0
+	github.com/immutable/imx-core-sdk-golang/imx/examples v0.0.0-00010101000000-000000000000
 )
 
 require (
@@ -20,6 +22,7 @@ require (
 	github.com/go-stack/stack v1.8.1 // indirect
 	github.com/google/uuid v1.2.0 // indirect
 	github.com/gorilla/websocket v1.4.2 // indirect
+	github.com/joho/godotenv v1.4.0 // indirect
 	github.com/rjeczalik/notify v0.9.1 // indirect
 	github.com/shirou/gopsutil v3.21.4-0.20210419000835-c7a38de76ee5+incompatible // indirect
 	github.com/tklauser/go-sysconf v0.3.5 // indirect
