+++ imx/examples/order/main.go
@@ -1,4 +1,4 @@
-package workflow
+package main
 
 import (
 	"context"
@@ -7,18 +7,30 @@ import (
 
 	"github.com/immutable/imx-core-sdk-golang/imx"
 	"github.com/immutable/imx-core-sdk-golang/imx/api"
+	"github.com/immutable/imx-core-sdk-golang/imx/examples/common"
 )
 
-const (
-	tokenAddress = "" // Provide your token address here, which is also contract address.
-	tokenID      = "" // Provide the token id being listed for sale.
-)
+func main() {
+	ctx, envs, c, l1signer, l2signer := common.CommonInitialise()
 
-func DemoOrdersWorkflow(ctx context.Context, c *imx.Client, amount imx.Wei, l1signer imx.L1Signer, l2signer imx.L2Signer) {
+	// Note: To list assets for sale on marketplace, you will need to have some assets added to account.
+
+	// Orders Demo
+	listingPriceInWei, err := strconv.ParseUint(envs["CREATE_ORDER_LISTING_PRICE_IN_WEI"], 10, 64)
+	if err != nil {
+		log.Panicf("error in converting ethAmountInWei from string to int: %v\n", err)
+	}
+
+	DemoOrdersWorkflow(ctx, c, envs["CREATE_ORDER_TOKEN_ID"],
+		envs["CREATE_ORDER_TOKEN_ADDRESS"], listingPriceInWei, l1signer, l2signer)
+}
+
+func DemoOrdersWorkflow(ctx context.Context, c *imx.Client, tokenID, tokenAddress string, amount imx.Wei, l1signer imx.L1Signer, l2signer imx.L2Signer) {
 	log.Println("-------------------------------------------------------")
-	log.Printf("Running %s", getCurrentFunctionName())
+	log.Printf("Running %s", common.GetCurrentFunctionName())
 
-	// The amount (listing price) should be in Wei for Eth tokens, see https://docs.starkware.co/starkex-v4/starkex-deep-dive/starkex-specific-concepts
+	// The amount (listing price) should be in Wei for Eth tokens,
+	// see https://docs.starkware.co/starkex-v4/starkex-deep-dive/starkex-specific-concepts
 	// and https://eth-converter.com/
 	ethAddress := l1signer.GetAddress()                         // Address of the user listing for sale.
 	sellToken := imx.SignableERC721Token(tokenID, tokenAddress) // NFT Token
@@ -39,7 +51,7 @@ func DemoOrdersWorkflow(ctx context.Context, c *imx.Client, amount imx.Wei, l1si
 		log.Panicf("error calling CreateOrder workflow: %v", err)
 	}
 
-	createOrderResponseStr, err := prettyStruct(createOrderResponse)
+	createOrderResponseStr, err := common.PrettyStruct(createOrderResponse)
 	if err != nil {
 		log.Panic(err)
 	}
@@ -55,12 +67,12 @@ func DemoOrdersWorkflow(ctx context.Context, c *imx.Client, amount imx.Wei, l1si
 		log.Panicf("error calling CreateOrder workflow: %v", err)
 	}
 
-	cancelOrderResponseStr, err := prettyStruct(cancelOrderResponse)
+	cancelOrderResponseStr, err := common.PrettyStruct(cancelOrderResponse)
 	if err != nil {
 		log.Panic(err)
 	}
 	log.Printf("CancelOrder response:\n%v\n", cancelOrderResponseStr)
 
-	log.Printf("Running %s completed.", getCurrentFunctionName())
+	log.Printf("Running %s completed.", common.GetCurrentFunctionName())
 	log.Println("-------------------------------------------------------")
 }
