--- imx/examples/workflow/utils.go
@@ -1,26 +1,0 @@
-package workflow
-
-import (
-	"encoding/json"
-	"os"
-	"runtime"
-)
-
-func prettyStruct(data interface{}) (string, error) {
-	val, err := json.MarshalIndent(data, "", "    ")
-	if err != nil {
-		return "", err
-	}
-	return string(val), nil
-}
-
-func getCurrentFunctionName() string {
-	counter, _, _, success := runtime.Caller(1)
-
-	if !success {
-		println("functionName: runtime.Caller: failed")
-		os.Exit(1)
-	}
-
-	return runtime.FuncForPC(counter).Name()
-}
