+++ imx/withdrawal.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/immutable/imx-core-sdk-golang/imx/api"
 	"github.com/immutable/imx-core-sdk-golang/imx/internal/convert"
@@ -139,9 +138,9 @@ func (c *Client) completeFungiblesWithdrawal(
 	assetType *big.Int,
 	overrides *bind.TransactOpts,
 ) (*types.Transaction, error) {
-	starkKey, err := hexutil.DecodeBig(starkKeyHex)
-	if err != nil {
-		return nil, fmt.Errorf("error converting StarkKeyHex to bigint: %s", starkKeyHex)
+	starkKey, ok := new(big.Int).SetString(starkKeyHex, 0)
+	if !ok {
+		return nil, fmt.Errorf("error converting starkKeyHex to bigint")
 	}
 
 	isRegistered, _ := c.registrationContract.IsRegistered(&bind.CallOpts{Context: ctx}, starkKey)
@@ -205,9 +204,9 @@ func (w *ERC721Withdrawal) withdrawMintedNft(
 		return nil, err
 	}
 
-	starkKey, err := hexutil.DecodeBig(starkKeyHex)
-	if err != nil {
-		return nil, fmt.Errorf("error converting StarkKeyHex to bigint: %s", starkKeyHex)
+	starkKey, ok := new(big.Int).SetString(starkKeyHex, 0)
+	if !ok {
+		return nil, fmt.Errorf("error converting starkKeyHex to bigint")
 	}
 
 	isRegistered, err := c.registrationContract.IsRegistered(&bind.CallOpts{Context: ctx}, starkKey)
@@ -300,9 +299,9 @@ func (w *ERC721Withdrawal) CompleteWithdrawal(
 		return nil, err
 	}
 
-	starkKey, err := hexutil.DecodeBig(starkKeyHex)
-	if err != nil {
-		return nil, fmt.Errorf("error converting StarkKeyHex to bigint: %s", starkKeyHex)
+	starkKey, ok := new(big.Int).SetString(starkKeyHex, 0)
+	if !ok {
+		return nil, fmt.Errorf("error converting starkKeyHex to bigint")
 	}
 
 	isRegistered, _ := c.registrationContract.IsRegistered(&bind.CallOpts{Context: ctx}, starkKey)
